languages/python/sqlalchemy-oso/requirements-test.txt
aiosqlite==0.17.0
pytest==7.1.2
pytest-asyncio==0.18.3
flask
flask_sqlalchemy
